Test image open/close: disabling volume opening test on invalid bootblock.
The ADFlib is not returning an error loading bootblock (it is read
just as a raw block and only the first 4 bytes indicating the type
of the volume's filesystem are checked). So the testing whether
opening a volume with invalid bootblock (checksum) fails (with
ADFlib's property ADF_PR_IGNORE_CHECKSUM_ERRORS unset) does not
pass (only the rootblock checksum is checked by the ADFlib).
Hence - disabling the test.
(It can eventually be enabled when used with a volume having
the rootblock with an invalid checksum).